requirements.txt
numpy==1.26.4
scikit-learn==1.7.2
pillow==12.0.0
face_recognition==1.3.0
tqdm==4.66.5